Router now handles URLs without locale or dialect.
This is achieved by setting the :localized and :translated config options to false. There then needs to be a default locale/dialect pair set in the backend — say global/en.
Also, when running the slice in dev mode, public routes are mounted at /public, so a localized route becomes /public/au/en. This is to stop the route colliding the the admin which is mounted at the root in dev.